WIP => add standalone mode for tests CI & run dev

+++ src/api/controllers/catalog.controller.ts
@@ -6,16 +6,26 @@ import { connectClient, disconnectClient, redisClient } from '../utils/redis/con
 import { FileProps } from '../utils/redis/types';
 import { purgeData, sendResponse } from '../middleware/validators/utils';
 import { getCurrentDateVersion } from '../utils/catalog';
+import fs from 'fs';
 
 export const addFileInCatalog = async (item: FileProps): Promise<{ status: number; message: string; data: Object }> => {
+    if (process.env.STANDALONE) {
+        const catalog = JSON.parse(fs.readFileSync('/tmp/standalone/catalog.json', 'utf-8')).data;
+        fs.writeFileSync('/tmp/standalone/catalog.json', JSON.stringify({ data: [ ...catalog, item ] }));
+        return {
+            status: 200,
+            message: `Item added with uuid`,
+            data: item
+        };
+    }
     try {
         await connectClient();
         const response = await addOneFile(item);
         await purgeData('catalog');
-        if (response.data && (!response.errors || response.errors.length === 0)) {
+        if (response.data && ( !response.errors || response.errors.length === 0 )) {
             return {
                 status: 200,
-                message: `Item added with uuid: ${response.data.uuid}`,
+                message: `Item added with uuid: ${ response.data.uuid }`,
                 data: response.data
             };
         }
@@ -24,11 +34,11 @@ export const addFileInCatalog = async (item: FileProps): Promise<{ status: numbe
             message: response.errors?.join(', ') || 'Unknown error',
             data: null
         };
-    } catch (err: unknown) {
-        logger.error(`Error adding file: ${err}`);
+    } catch ( err: unknown ) {
+        logger.error(`Error adding file: ${ err }`);
         return {
             status: 500,
-            message: `Error adding file: ${err}`,
+            message: `Error adding file: ${ err }`,
             data: null
         };
     } finally {
@@ -37,17 +47,20 @@ export const addFileInCatalog = async (item: FileProps): Promise<{ status: numbe
 };
 
 export const getFiles = async (req: Request, res: Response) => {
+    if (process.env.STANDALONE) {
+        return res.status(200).send(JSON.parse(fs.readFileSync('/tmp/standalone/catalog.json', 'utf-8')).data);
+    }
     try {
         await connectClient();
         const response = await getAllFiles();
-        if (response.data && (!response.errors || response.errors.length === 0)) {
+        if (response.data && ( !response.errors || response.errors.length === 0 )) {
             return res.status(200).send(response.data);
         }
         return res.status(500).send(response.errors);
-    } catch (err: unknown) {
-        logger.error(`Error getting files: ${err}`);
+    } catch ( err: unknown ) {
+        logger.error(`Error getting files: ${ err }`);
         return res.status(500).send({
-            message: `Error getting files: ${err}`
+            message: `Error getting files: ${ err }`
         });
     } finally {
         await disconnectClient();
@@ -55,21 +68,32 @@ export const getFiles = async (req: Request, res: Response) => {
 };
 
 export const getFile = async (req: Request, res: Response) => {
-    try {
-        await connectClient();
-        const response = await getOneFile(req.params.id);
-        if (response.data && (!response.errors || response.errors.length === 0)) {
-            return res.status(200).json({ data: [response.data] });
+    if (process.env.STANDALONE) {
+        const catalog = JSON.parse(fs.readFileSync('/tmp/standalone/catalog.json', 'utf-8')).data;
+        const item = catalog.find(item => item.uuid === req.params.id);
+        if (item) {
+            return res.status(200).send({ data: [ item ] });
         }
         return res.status(404).json({
             data: null,
-            errors: [`Unable to find file with id ${req.params.id} => ${response.errors?.join(', ')}`]
+            errors: [ `Unable to find file with id ${ req.params.id }` ]
         });
-    } catch (err: unknown) {
-        logger.error(`Error getting file: ${err}`);
+    }
+    try {
+        await connectClient();
+        const response = await getOneFile(req.params.id);
+        if (response.data && ( !response.errors || response.errors.length === 0 )) {
+            return res.status(200).json({ data: [ response.data ] });
+        }
+        return res.status(404).json({
+            data: null,
+            errors: [ `Unable to find file with id ${ req.params.id } => ${ response.errors?.join(', ') }` ]
+        });
+    } catch ( err: unknown ) {
+        logger.error(`Error getting file: ${ err }`);
         return res.status(500).send({
             data: null,
-            errors: [`Error getting file: ${err}`]
+            errors: [ `Error getting file: ${ err }` ]
         });
     } finally {
         await disconnectClient();
@@ -77,6 +101,15 @@ export const getFile = async (req: Request, res: Response) => {
 };
 
 export const updateFileInCatalog = async (uuid: string, itemToUpdate: FileProps): Promise<any> => {
+    if (process.env.STANDALONE) {
+        const catalog = JSON.parse(fs.readFileSync('/tmp/standalone/catalog.json', 'utf-8')).data;
+        const updatedCatalog = catalog.map(item => {
+            if (item.uuid === uuid) {return { ...item, ...itemToUpdate };}
+            return item;
+        });
+        fs.writeFileSync('/tmp/standalone/catalog.json', JSON.stringify({ data: updatedCatalog }));
+        return { data: itemToUpdate };
+    }
     await connectClient();
     const updateItem = await updateOneFile(uuid, itemToUpdate);
     await disconnectClient();
@@ -85,33 +118,46 @@ export const updateFileInCatalog = async (uuid: string, itemToUpdate: FileProps)
 };
 
 export const deleteFileFromCatalog = async (uniqueName: string): Promise<{ status: number; message: string }> => {
+    const catalog = await getCatalog();
+    if (process.env.STANDALONE) {
+        const updatedCatalog = catalog.data.filter(item => item.unique_name !== uniqueName);
+        fs.writeFileSync('/tmp/standalone/catalog.json', JSON.stringify({ data: updatedCatalog }));
+        return {
+            status: 200,
+            message: `Successfully deleted ${ uniqueName }`
+        };
+    }
     try {
-        const catalog = await getCatalog();
+
         const itemFound = catalog.data.find((item) => item.unique_name === uniqueName);
         if (!itemFound) {
-            return { status: 404, message: `Item not found: ${uniqueName}` };
+            return { status: 404, message: `Item not found: ${ uniqueName }` };
         }
 
         await connectClient();
         await redisClient.del(itemFound.uuid);
         await disconnectClient();
         await purgeData('catalog');
-        return { status: 200, message: `Successfully deleted ${uniqueName}` };
-    } catch (err: unknown) {
-        logger.error(`Error deleting file: ${err}`);
+        return { status: 200, message: `Successfully deleted ${ uniqueName }` };
+    } catch ( err: unknown ) {
+        logger.error(`Error deleting file: ${ err }`);
         return {
             status: 500,
-            message: `Error deleting file: ${(err as Error).message}`
+            message: `Error deleting file: ${ ( err as Error ).message }`
         };
     }
 };
 
 export const deleteCatalog = async (req, res) => {
+    if (process.env.STANDALONE) {
+        fs.writeFileSync('/tmp/standalone/catalog.json', JSON.stringify({ data: [] }));
+        return res.status(200).send('ALL GOOD');
+    }
     await connectClient();
     const response = await getAllFiles();
     if (response.data) {
         await purgeData('catalog');
-        for (const item of response.data) {
+        for ( const item of response.data ) {
             await deleteFileFromCatalog(item.unique_name);
         }
     }
@@ -122,16 +168,16 @@ export const deleteCatalog = async (req, res) => {
 export const createDump = async (req: Request, res: Response) => {
     const { data: catalog } = await getCatalog();
     const fileVersion = getCurrentDateVersion();
-    const filePath = `${app.locals.PREFIXED_CATALOG}/${fileVersion}.json`;
+    const filePath = `${ app.locals.PREFIXED_CATALOG }/${ fileVersion }.json`;
     if (!fileVersion) {
         return sendResponse({
             res,
             status: 400,
-            data: ['Error generating dump.json from Redis client']
+            data: [ 'Error generating dump.json from Redis client' ]
         });
     }
     if (fileVersion) {
-        const postBackupFileJson = await fetch(`${app.locals.PREFIXED_API_URL}/backup?filepath=${filePath}`, {
+        const postBackupFileJson = await fetch(`${ app.locals.PREFIXED_API_URL }/backup?filepath=${ filePath }`, {
             method: 'POST',
             headers: { 'Content-Type': 'application/json' },
             body: JSON.stringify(catalog)
@@ -140,18 +186,18 @@ export const createDump = async (req: Request, res: Response) => {
             return sendResponse({
                 res,
                 status: 400,
-                data: ['Failed to upload JSON  in backup']
+                data: [ 'Failed to upload JSON  in backup' ]
             });
         }
         return sendResponse({
             res,
             status: 200,
-            data: ['DUMP backup successfully']
+            data: [ 'DUMP backup successfully' ]
         });
     }
     return sendResponse({
         res,
         status: 200,
-        data: ['Successfully generated dump.rdb from Redis client']
+        data: [ 'Successfully generated dump.rdb from Redis client' ]
     });
 };
